refactor: Remove unnecessary layer of abstraction for page components
